backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
google-generativeai>=0.8.0
google-genai>=0.3.0
pdfminer.six>=20231228
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
gunicorn>=21.0.0
pydantic>=2.0.0
tqdm>=4.66.0